P4ADEV-3023 code review

+++ src/main/java/it/gov/pagopa/pu/migration/dto/debtpositiontypeorgoperator/DebtPositionTypeOrgOperatorMigrationFileResult.java
@@ -14,7 +14,7 @@ import java.util.List;
 @ToString(callSuper = true)
 public class DebtPositionTypeOrgOperatorMigrationFileResult extends MigrationFileResult {
   private Long organizationId;
-  private int numTotalRows;
-  private int numCorrectlyProcessedRows;
+  private Long numTotalRows;
+  private Long numCorrectlyProcessedRows;
   private List<DebtPositionTypeOrgOperatorErrorDTO> errorList;
 }
